# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me/invitations": {
      "get": {
        "operationId": "editorsInvitations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invitation"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Invitations waiting for this account's answer.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/invitations/{inviteId}/accept": {
      "post": {
        "operationId": "editorsAccept",
        "parameters": [
          {
            "name": "inviteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Accept one, taking a seat on that group's page.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/invitations/{inviteId}/decline": {
      "post": {
        "operationId": "editorsDecline",
        "parameters": [
          {
            "name": "inviteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Decline one. The owner can ask again later.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/groups/{pageId}/editors": {
      "get": {
        "operationId": "editorsList",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupEditors"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Who runs this group: seats taken, then invitations waiting. Owners only.",
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "editorsInvite",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Editor"
                }
              }
            }
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Ask the person at a vrc.page name to help run this group. Owners only.",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/groups/{pageId}/editors/{id}": {
      "delete": {
        "operationId": "editorsDismiss",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error, as RFC 9457 Problem Details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "summary": "Take a row off that list: an invitation taken back, an editor removed, or\nan editor leaving, which anybody may do for their own seat.",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "vrc.page API",
    "description": "The backend for vrc.page. Errors use RFC 9457 Problem Details (application/problem+json), and every response carries an X-Request-Id header.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Invitation": {
        "type": "object",
        "properties": {
          "groupSlug": {
            "type": "string",
            "nullable": true
          },
          "groupIconUrl": {
            "type": "string",
            "nullable": true
          },
          "memberCount": {
            "type": "integer",
            "nullable": true
          },
          "invitedBy": {
            "type": "string",
            "nullable": true,
            "description": "Who asked, by their VRChat name."
          },
          "id": {
            "type": "string"
          },
          "groupName": {
            "type": "string",
            "description": "The group being invited to."
          },
          "sentAt": {
            "type": "string",
            "description": "ISO 8601."
          }
        },
        "required": [
          "groupSlug",
          "groupIconUrl",
          "memberCount",
          "invitedBy",
          "id",
          "groupName",
          "sentAt"
        ]
      },
      "Editor": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "description": "Their VRChat name, or null if they have disconnected VRChat since."
          },
          "slug": {
            "type": "string",
            "nullable": true,
            "description": "Their own page, if it has a name."
          },
          "iconUrl": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "description": "This account, as this group's owner may refer to it. Opaque: it is only\never used to take the seat away again."
          },
          "since": {
            "type": "string",
            "description": "When they took the seat, or when they were asked: ISO 8601."
          },
          "pending": {
            "type": "boolean",
            "description": "An invitation nobody has answered yet."
          }
        },
        "required": [
          "name",
          "slug",
          "iconUrl",
          "id",
          "since",
          "pending"
        ]
      },
      "GroupEditors": {
        "type": "object",
        "properties": {
          "max": {
            "type": "integer",
            "description": "Seats and waiting invitations together may not pass this."
          },
          "editors": {
            "description": "Seats taken, then invitations still waiting.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Editor"
            }
          }
        },
        "required": [
          "max",
          "editors"
        ]
      },
      "InviteRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Their vrc.page name, as in vrc.page/<name>."
          }
        },
        "required": [
          "name"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "The HTTP status code.",
            "example": 404
          },
          "retryAfter": {
            "type": "integer",
            "description": "Seconds to wait before trying again, for a cooldown."
          },
          "at": {
            "type": "integer",
            "description": "Which item of a submitted list the problem is about, counting from 0."
          },
          "type": {
            "type": "string",
            "description": "A URI naming the kind of problem: \"about:blank\" when the status says it\nall, otherwise https://vrc.page/problems/<code>, with a code from\nsrc/common/problem.ts."
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the kind of problem."
          },
          "detail": {
            "type": "string",
            "description": "What went wrong this time, in words a person can act on."
          },
          "instance": {
            "type": "string",
            "description": "The path of the request that failed."
          },
          "requestId": {
            "type": "string",
            "description": "Same as the X-Request-Id response header and the request's audit log entries."
          }
        },
        "required": [
          "status",
          "type",
          "title",
          "requestId"
        ]
      }
    }
  }
}